requirements.txt
asgiref==3.4.1
Django==3.2.7
django-tempus-dominus==5.1.2.16
django-widget-tweaks==1.4.8
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
